chore: 패키지 구조 변경
- 레이어 구조에서 모듈 구조로 변경하였습니다.
