--- BE/src/main/java/team24/issuetracker/domain/Comment.java
@@ -1,36 +1,0 @@
-package team24.issuetracker.domain;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-
-public class Comment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Issue issue;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User writer;
-    private String content;
-    private LocalDateTime dateTime;
-    private boolean isIssueWriter;
-}
