+++ BE/src/main/java/team24/issuetracker/issue/domain/Image.java
@@ -1,0 +1,29 @@
+package team24.issuetracker.issue.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import team24.issuetracker.issue.domain.Issue;
+
+@Getter
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Image {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String imagePath;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Issue issue;
+}
