+++ BE/src/main/java/team24/issuetracker/issue/domain/dto/IssueListMilestoneResponse.java
@@ -1,0 +1,16 @@
+package team24.issuetracker.issue.domain.dto;
+
+import lombok.Getter;
+import team24.issuetracker.milestone.domain.Milestone;
+
+@Getter
+public class IssueListMilestoneResponse {
+
+	private final Long id;
+	private final String title;
+
+	public IssueListMilestoneResponse(Milestone milestone) {
+		this.id = milestone.getId();
+		this.title = milestone.getTitle();
+	}
+}
