+++ BE/src/main/java/team24/issuetracker/issue/domain/reference/IssueLabel.java
@@ -1,0 +1,32 @@
+package team24.issuetracker.issue.domain.reference;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import team24.issuetracker.label.domain.Label;
+import team24.issuetracker.issue.domain.Issue;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class IssueLabel {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Label label;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Issue issue;
+}
