+++ BE/src/main/java/team24/issuetracker/issue/domain/reference/IssueUser.java
@@ -1,0 +1,32 @@
+package team24.issuetracker.issue.domain.reference;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import team24.issuetracker.common.domain.User;
+import team24.issuetracker.issue.domain.Issue;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class IssueUser {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Issue issue;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private User user;
+}
