+++ BE/src/main/java/team24/issuetracker/milestone/domain/Milestone.java
@@ -1,0 +1,33 @@
+package team24.issuetracker.milestone.domain;
+
+import java.time.LocalDate;
+import java.util.List;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import team24.issuetracker.issue.domain.Issue;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Milestone {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String title;
+    private String description;
+    private LocalDate dueDate;
+    private boolean isDeleted;
+
+    @OneToMany(mappedBy = "milestone")
+    private List<Issue> issues;
+}
