--- BE/src/main/java/team24/issuetracker/web/config/SwaggerConfig.java
@@ -1,39 +1,0 @@
-package team24.issuetracker.web.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-@Configuration
-public class SwaggerConfig {
-	@Bean
-	public Docket api() {
-		return new Docket(DocumentationType.SWAGGER_2)
-			.select()
-			.apis(RequestHandlerSelectors.any())
-			.paths(PathSelectors.any())
-			.build()
-			.groupName("API 1.0.0")
-			.pathMapping("/")
-			.apiInfo(apiInfo())
-			.useDefaultResponseMessages(false);
-	}
-	private ApiInfo apiInfo() {
-		return new ApiInfoBuilder()
-			.title("fantoo_api_test 프로젝트")
-			.description("API 호출 테스트용도.")
-			.version("1.0.0")
-			.termsOfServiceUrl("")
-			.license("")
-			.licenseUrl("")
-			.build();
-	}
-}
